backend/requirements.txt
python-socketio==5.6.0
aiohttp==3.8.3
aiosignal==1.3.1
argparse-addons==0.8.0
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.0
bitstruct==8.15.1
can==0.0.0
cantools==38.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crccheck==1.3.0
diskcache==5.4.0
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
packaging==23.0
pyserial==3.5
python-can==4.1.0
python-engineio==4.3.2
pywin32==305
requests==2.28.2
simple-websocket==0.9.0
six==1.16.0
textparser==0.24.0
typing_extensions==4.4.0
urllib3==1.26.14
websocket-client==1.5.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
digi-xbee==1.4.1
reverb-serial==2.0.1